+++ crates/rules/src/roles/factor_levels/factor_instance_level/mod.rs
@@ -1,0 +1,11 @@
+mod confirmation_role_with_factor_instances;
+mod general_role_with_hierarchical_deterministic_factor_instances;
+mod primary_role_with_factor_instances;
+mod recovery_role_with_factor_instances;
+mod role_with_factor_instances;
+
+pub(crate) use confirmation_role_with_factor_instances::*;
+pub use general_role_with_hierarchical_deterministic_factor_instances::*;
+pub(crate) use primary_role_with_factor_instances::*;
+pub(crate) use recovery_role_with_factor_instances::*;
+pub(crate) use role_with_factor_instances::*;
