+++ crates/rules/src/roles/factor_levels/factor_source_id_level/roles_with_factor_ids.rs
@@ -1,0 +1,3 @@
+use crate::prelude::*;
+
+pub type RoleWithFactorSourceIds<const R: u8> = AbstractBuiltRoleWithFactor<R, FactorSourceID>;
