+++ crates/rules/src/roles/mod.rs
@@ -1,13 +1,7 @@
 mod abstract_role_builder_or_built;
 mod builder;
-mod general_role_with_hierarchical_deterministic_factor_instances;
-mod role_with_factor_instances;
-mod roles_with_factor_ids;
-mod roles_with_factor_sources;
+mod factor_levels;
 
 pub(crate) use abstract_role_builder_or_built::*;
 pub use builder::*;
-pub use general_role_with_hierarchical_deterministic_factor_instances::*;
-pub(crate) use role_with_factor_instances::*;
-pub use roles_with_factor_ids::*;
-pub(crate) use roles_with_factor_sources::*;
+pub use factor_levels::*;
